feat: add zustand and store
