+++ src/feature/comment/ui/Comments.tsx
@@ -1,40 +1,33 @@
 import { Button, HighlightText } from "../../../shared/ui"
 import { Edit2, Plus, ThumbsUp, Trash2 } from "lucide-react"
 import { Comment } from "../../../entities"
-import { NewComment } from "../type"
+import { useCommentStore } from "../model/store"
+import { useSearchQueryStore, useSelectedPostStore } from "../../post/model/store"
 
-export const Comments = ({
-  postId,
-  searchQuery,
-  setNewComment,
-  setShowAddCommentDialog,
-  comments,
-  setComments,
-  setSelectedComment,
-  setShowEditCommentDialog,
-}: {
-  postId: number
-  searchQuery: string
-  setNewComment: React.Dispatch<React.SetStateAction<NewComment>>
-  setShowAddCommentDialog: React.Dispatch<React.SetStateAction<boolean>>
-  comments: {
-    [key: number]: Comment[]
-  }
-  setSelectedComment: React.Dispatch<React.SetStateAction<Comment | null>>
-  setShowEditCommentDialog: React.Dispatch<React.SetStateAction<boolean>>
-  setComments: React.Dispatch<
-    React.SetStateAction<{
-      [key: number]: Comment[]
-    }>
-  >
-}) => {
+export const Comments = () => {
+  const { searchQuery } = useSearchQueryStore()
+  const { selectedPost } = useSelectedPostStore()
+  const {
+    comments,
+    setComments,
+    setNewComment,
+    setShowAddCommentDialog,
+    setShowEditCommentDialog,
+    setSelectedComment,
+  } = useCommentStore()
+  console.log(1111)
+
+  console.log(comments)
+
+  const postId = selectedPost.id
   // 댓글 좋아요
   const likeComment = async (id: number, postId: number) => {
+    const filtredComments = comments[postId]
     try {
       const response = await fetch(`/api/comments/${id}`, {
         method: "PATCH",
         headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ likes: comments[postId].find((c) => c.id === id).likes + 1 }),
+        body: JSON.stringify({ likes: filtredComments.find((comment) => (comment as Comment).id === id)!.likes + 1 }),
       })
       const data = await response.json()
       setComments((prev) => ({
@@ -63,49 +56,49 @@ export const Comments = ({
     }
   }
 
+  const handleAddComment = () => {
+    setNewComment((prev) => ({ ...prev, postId }))
+    setShowAddCommentDialog(true)
+  }
+
   return (
     <div className="mt-2">
       <div className="flex items-center justify-between mb-2">
         <h3 className="text-sm font-semibold">댓글</h3>
-        <Button
-          size="sm"
-          onClick={() => {
-            setNewComment((prev) => ({ ...prev, postId }))
-            setShowAddCommentDialog(true)
-          }}
-        >
+        <Button size="sm" onClick={handleAddComment}>
           <Plus className="w-3 h-3 mr-1" />
           댓글 추가
         </Button>
       </div>
       <div className="space-y-1">
-        {comments[postId]?.map((comment) => (
-          <div key={comment.id} className="flex items-center justify-between text-sm border-b pb-1">
-            <div className="flex items-center space-x-2 overflow-hidden">
-              <span className="font-medium truncate">{comment.user.username}:</span>
-              <span className="truncate">{HighlightText(comment.body, searchQuery)}</span>
+        {Object.keys(comments).length > 0 &&
+          comments[postId].map((comment) => (
+            <div key={comment.id} className="flex items-center justify-between text-sm border-b pb-1">
+              <div className="flex items-center space-x-2 overflow-hidden">
+                <span className="font-medium truncate">{comment.user.username}:</span>
+                <span className="truncate">{HighlightText(comment.body, searchQuery)}</span>
+              </div>
+              <div className="flex items-center space-x-1">
+                <Button variant="ghost" size="sm" onClick={() => likeComment(comment.id, postId)}>
+                  <ThumbsUp className="w-3 h-3" />
+                  <span className="ml-1 text-xs">{comment.likes}</span>
+                </Button>
+                <Button
+                  variant="ghost"
+                  size="sm"
+                  onClick={() => {
+                    setSelectedComment(comment)
+                    setShowEditCommentDialog(true)
+                  }}
+                >
+                  <Edit2 className="w-3 h-3" />
+                </Button>
+                <Button variant="ghost" size="sm" onClick={() => deleteComment(comment.id, postId)}>
+                  <Trash2 className="w-3 h-3" />
+                </Button>
+              </div>
             </div>
-            <div className="flex items-center space-x-1">
-              <Button variant="ghost" size="sm" onClick={() => likeComment(comment.id, postId)}>
-                <ThumbsUp className="w-3 h-3" />
-                <span className="ml-1 text-xs">{comment.likes}</span>
-              </Button>
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => {
-                  setSelectedComment(comment)
-                  setShowEditCommentDialog(true)
-                }}
-              >
-                <Edit2 className="w-3 h-3" />
-              </Button>
-              <Button variant="ghost" size="sm" onClick={() => deleteComment(comment.id, postId)}>
-                <Trash2 className="w-3 h-3" />
-              </Button>
-            </div>
-          </div>
-        ))}
+          ))}
       </div>
     </div>
   )
