+++ src/shared/ui/dialog/dialog/type.ts
@@ -1,6 +1,6 @@
 export interface DialogProps {
   open: boolean
   title: string | React.ReactNode
-  handleChange: React.Dispatch<React.SetStateAction<boolean>>
+  handleChange: (open: boolean) => void
   children: React.ReactNode
 }
